fixed some issues, tests still wont pass

+++ router/storyrouter.js
@@ -12,8 +12,8 @@ module.exports = function(router){
     //   on failer: send error back
     if(!req.body.name || !req.body.text){
       let err = new Error('Did not enter text');
-      console.error(err);
       res.statusCode = 400; // bad request
+      console.error(err);
       res.end();
       return;
     }
@@ -37,6 +37,11 @@ module.exports = function(router){
   router.get('/api/story', function(req, res){
     let id = req.url.query.id;
     // TODO: put logic right here for a 400 if no id
+    if(!id) {
+      res.statusCode = 400;
+      res.end();
+      return;
+    }
     storage.getItem('story', id)
     .then(note => {
       res.setHeader('Content-Type', 'application/json');
@@ -45,40 +50,37 @@ module.exports = function(router){
     })
     // TODO:  put logic in here for a 404 if getItem didnt find a note
     .catch(err => {
-      if (typeof err === Error) {
-        if (err.status === 404) {
-          res.statusCode = 404;
-          res.end();
-          return;
-        }
-      }
-      // make better errors
+      res.statusCode = err.status;
       console.error(err);
-      res.statusCode = 500;
       res.end();
+      return;
     });
+      // make better errors
+    res.statusCode = 500;
+    res.end();
   });
   //**********************DELETE*************************************
   router.delete('/api/story', function (req, res){
     let id =req.url.query.id;
-    storage.deleteItem('story' , id)
+    if (!id) {
+      res.statusCode = 400;
+      res.end();
+      return;
+    }
+    storage.deleteItem('story', id)
     .then(() => {
       res.setHeader('Content-Type', 'application/json');
-      res.write(JSON.stringify(id));
+      res.statusCode = 204;
       res.end();
     })
     .catch(err => {
-      if (typeof err === Error) {
-        if (err.status === 404) {
-          res.statusCode = 404;
-          res.end();
-          return;
-        }
-      }
-      // make better errors
+      res.statusCode = err.status;
       console.error(err);
-      res.statusCode = 500;
       res.end();
+      return;
     });
+      // make better errors
+    res.statusCode = 500;
+    res.end();
   });
 };
